moved formatGrafanaDependency to helpers, added <9 version handling
