+++ public/app/features/plugins/admin/components/VersionList.tsx
@@ -7,11 +7,10 @@ import { Trans, t } from '@grafana/i18n';
 import { config } from '@grafana/runtime';
 import { useStyles2, Badge } from '@grafana/ui';
 
-import { getLatestCompatibleVersion, shouldDisablePluginInstall } from '../helpers';
+import { formatGrafanaDependency, getLatestCompatibleVersion, shouldDisablePluginInstall } from '../helpers';
 import { type CatalogPlugin, PluginUpdateStrategy, type Version } from '../types';
 
 import { VersionInstallButton } from './VersionInstallButton';
-import { formatGrafanaDependency } from './formatGrafanaDependency';
 
 interface Props {
   plugin: CatalogPlugin;
